Renames packages to org.fcrepo.spec.testsuite.
