--- src/main/java/com/ibr/fedora/testsuite/SetUpSuite.java
@@ -1,39 +1,0 @@
-/**
- * @author Fernando Cardoza
- */
-/*
- * Licensed to DuraSpace under one or more contributor license agreements.
- * See the NOTICE file distributed with this work for additional information
- * regarding copyright ownership.
- *
- * DuraSpace licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except in
- * compliance with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.ibr.fedora.testsuite;
-
-import java.io.FileNotFoundException;
-
-import com.ibr.fedora.TestSuiteGlobals;
-import org.testng.annotations.BeforeSuite;
-import org.testng.annotations.Listeners;
-
-@Listeners({com.ibr.fedora.report.HtmlReporter.class, com.ibr.fedora.report.EarlReporter.class})
-public class SetUpSuite {
-
-    /**
-     * @throws FileNotFoundException
-     */
-    @BeforeSuite
-    public void setUp() throws FileNotFoundException {
-        TestSuiteGlobals.resetFile();
-    }
-}
