+++ src/main/java/org/fcrepo/spec/testsuite/tests/HttpDelete.java
@@ -1,30 +1,24 @@
+/*
+ *
+ *  * The contents of this file are subject to the license and copyright
+ *  * detailed in the LICENSE and NOTICE files at the root of the source
+ *  * tree and available online at
+ *  *
+ *  *     http://duracloud.org/license/
+ *
+ */
+
 /**
  * @author Jorge Abrego, Fernando Cardoza
  */
-/*
- * Licensed to DuraSpace under one or more contributor license agreements.
- * See the NOTICE file distributed with this work for additional information
- * regarding copyright ownership.
- *
- * DuraSpace licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except in
- * compliance with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.ibr.fedora.testsuite;
+
+package org.fcrepo.spec.testsuite.tests;
 
 import java.io.FileNotFoundException;
 import java.io.PrintStream;
 
-import com.ibr.fedora.TestSuiteGlobals;
-import com.ibr.fedora.TestsLabels;
+import org.fcrepo.spec.testsuite.TestSuiteGlobals;
+import org.fcrepo.spec.testsuite.TestsLabels;
 import io.restassured.RestAssured;
 import io.restassured.config.LogConfig;
 import io.restassured.http.Header;
